% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{backend_future_lapply_staged}
\alias{backend_future_lapply_staged}
\title{Staged parallelism for the drake R package}
\usage{
backend_future_lapply_staged(config)
}
\arguments{
\item{config}{a \code{drake_config()} object}
}
\description{
With staged parallelism,
\code{drake} partitions the dependency graph into stages of
conditionally independent targets and processes each
stage with semi-transient parallel workers.
This functionality is already deprecated,
and it will be removed at some point later on.
}
\examples{
# See <https://github.com/wlandau/drake.future.lapply.staged/blob/master/README.md> # nolint
# for examples.
}
