% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{drake.future.lapply.staged}
\alias{drake.future.lapply.staged}
\alias{drake.future.lapply.staged-package}
\title{Staged parallelism for the drake R package.}
\description{
This package implements future-based staged parallelism
for the drake R package. The functionality is already deprecated,
and it will fall into disuse at some point.
}
\examples{
# See <https://github.com/wlandau/drake.future.lapply.staged/blob/master/README.md> # nolint
# for examples.
}
\references{
\url{https://github.com/wlandau/drake.future.lapply.staged}
}
\author{
William Michael Landau \email{will.landau@gmail.com}
}
